Release v1.5.3: Add keyboard shortcuts for drawing modes
- Add Alt+1-4 shortcuts for instant mode switching:
  * Alt+1: Freehand drawing mode
  * Alt+2: Rectangle drawing mode
  * Alt+3: Circle drawing mode
  * Alt+4: Arrow drawing mode
- Implement global keyboard detection with Windows API
- Add persistent settings for shortcut-triggered mode changes
- Create new v1.5.3 build scripts for all platforms
- Update changelog and version throughout codebase

+++ GrafTrail/overlay.py
@@ -656,13 +656,13 @@ class Overlay(QtWidgets.QWidget):
         glow.setAlpha(int(fade * GLOW_ALPHA_MAX))
         painter.setPen(QtCore.Qt.NoPen)              # No outline
         painter.setBrush(QtGui.QBrush(glow))         # Solid fill
-        painter.drawEllipse(center, self.glow_pen.width()/2, self.glow_pen.width()/2)
+        painter.drawEllipse(center, int(self.glow_pen.width()/2), int(self.glow_pen.width()/2))
 
         # Draw core circle (inner, more opaque)
         core = QtGui.QColor(col)
         core.setAlpha(int(fade * CORE_ALPHA_MAX))
         painter.setBrush(QtGui.QBrush(core))         # Solid fill
-        painter.drawEllipse(center, self.core_pen.width()/2, self.core_pen.width()/2)
+        painter.drawEllipse(center, int(self.core_pen.width()/2), int(self.core_pen.width()/2))
 
     # ===================================================================
     # RENDERING
